With clauses
* change select_statement to select_type and add select_expr
* With clauses, including multiple and nested with clauses
* fix sub query code for styling to be consistent
