+++ tests/snapshots/arg_item/020_arg_item_with_date_interval
@@ -13,9 +13,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree(
                     "select_list",
                     [
