+++ tests/snapshots/from_item/010_single_table
@@ -13,9 +13,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree("select_list", [Tree("select_item_unaliased", [Token("CNAME", "field")])]),
                 Token("FROM", "from"),
                 Tree("from_clause", [Tree("simple_table_name", [Token("CNAME", "table")])]),
