+++ tests/snapshots/from_item/100_sub_query
@@ -19,9 +19,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree("select_list", [Tree("select_item_unaliased", [Token("CNAME", "field")])]),
                 Token("FROM", "from"),
                 Tree(
@@ -30,10 +30,11 @@ Tree(
                         Tree(
                             "sub_query_expr",
                             [
+                                Token("LPAR", "("),
                                 Tree(
-                                    "query_expr",
+                                    "select_expr",
                                     [
-                                        Tree("select_statement", [Token("SELECT", "select")]),
+                                        Tree("select_type", [Token("SELECT", "select")]),
                                         Tree(
                                             "select_list",
                                             [
@@ -44,7 +45,8 @@ Tree(
                                         Token("FROM", "from"),
                                         Tree("from_clause", [Tree("simple_table_name", [Token("CNAME", "table")])]),
                                     ],
-                                )
+                                ),
+                                Token("RPAR", ")"),
                             ],
                         )
                     ],
