+++ tests/snapshots/limit/010_basic_limit
@@ -14,9 +14,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree("select_list", [Tree("select_item_unaliased", [Token("STAR", "*")])]),
                 Token("FROM", "from"),
                 Tree("from_clause", [Tree("simple_table_name", [Token("CNAME", "table")])]),
