+++ tests/snapshots/order_by/030_single_order_by_func
@@ -15,9 +15,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree(
                     "select_list",
                     [
