+++ tests/snapshots/order_by/100_order_by_asc
@@ -16,9 +16,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree(
                     "select_list",
                     [
