+++ tests/snapshots/select_item/040_select_item_explicit_alias
@@ -13,9 +13,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree(
                     "select_list",
                     [Tree("select_item_aliased", [Token("CNAME", "field"), Token("CNAME", "new_field_name")])],
