+++ tests/snapshots/select_item/100_simple_table_ref
@@ -13,9 +13,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree(
                     "select_list",
                     [
