+++ tests/snapshots/where_clause/210_multiple_binary_bool_expression
@@ -17,9 +17,9 @@ Tree(
     "query_file",
     [
         Tree(
-            "query_expr",
+            "select_expr",
             [
-                Tree("select_statement", [Token("SELECT", "select")]),
+                Tree("select_type", [Token("SELECT", "select")]),
                 Tree("select_list", [Tree("select_item_unaliased", [Token("CNAME", "field")])]),
                 Token("FROM", "from"),
                 Tree(
