+++ tests/snapshots/with_clause/010_basic_with_clause
@@ -1,0 +1,61 @@
+with cte as (
+	select * from table_a
+)
+select * from cte
+
+---
+
+WITH
+
+cte AS (
+    SELECT
+        *,
+    FROM
+        table_a
+)
+
+SELECT
+    *,
+FROM
+    cte
+
+---
+
+Tree(
+    "query_file",
+    [
+        Tree(
+            "query_expr",
+            [
+                Token("WITH", "with"),
+                Tree(
+                    "with_clause",
+                    [
+                        Token("CNAME", "cte"),
+                        Token("AS", "as"),
+                        Token("LPAR", "("),
+                        Tree(
+                            "select_expr",
+                            [
+                                Tree("select_type", [Token("SELECT", "select")]),
+                                Tree("select_list", [Tree("select_item_unaliased", [Token("STAR", "*")])]),
+                                Token("FROM", "from"),
+                                Tree("from_clause", [Tree("simple_table_name", [Token("CNAME", "table_a")])]),
+                            ],
+                        ),
+                        Token("RPAR", ")"),
+                    ],
+                ),
+                Tree(
+                    "select_expr",
+                    [
+                        Tree("select_type", [Token("SELECT", "select")]),
+                        Tree("select_list", [Tree("select_item_unaliased", [Token("STAR", "*")])]),
+                        Token("FROM", "from"),
+                        Tree("from_clause", [Tree("simple_table_name", [Token("CNAME", "cte")])]),
+                    ],
+                ),
+            ],
+        )
+    ],
+)
